requirement.txt
mariadb==1.1.4
mysql-connector-python==8.0.31
protobuf==3.20.1
